requirements.txt
fastapi
Werkzeug
requests
ffmpeg-python
git+https://github.com/openai/whisper.git
gunicorn
APScheduler
srt
numpy
torch
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
psutil
Pillow
boto3
uvicorn
aiohttp
wget
python-multipart
rembg